Exercise8/requirements.txt
google-cloud-datastore
requests
